Debug scaffolding, improved interpolation
When interpolating the subrender location, the mode is now explicitly
set so it does not default to nearest.

Added code to terminate sampling on after first subrender to assist in
visualize the correctness of the operation.

Added code to allow more convenient setting of the weighting applied to
the subrender.